k8s: move to client-go package
This makes a minimal set of changes to migrate to the client-go package
over. Most importantly, this gets rid of the abomination that was
'apihost' and replaces it with InClusterConfig.

This is not intended to change any functionality. This includes not
changing in any significant way the endpoints that are hit, when, and
why.